update method for getting end of day data

+++ src/swe/engine/HistoryState.java
@@ -5,9 +5,9 @@ import java.util.Calendar;
 final public class HistoryState {
     final public TickIdentifier tickIdentifier;
     final public MarketStatus status;
-    final public int shareIndex;
+    final public Float shareIndex;
 
-    public HistoryState(TickIdentifier tickIdentifier, MarketStatus status, int shareIndex) {
+    public HistoryState(TickIdentifier tickIdentifier, MarketStatus status, Float shareIndex) {
         this.tickIdentifier = tickIdentifier;
         this.status = status;
         this.shareIndex = shareIndex;
